refactor: move files under pkg
