--- Gopkg.toml
@@ -1,50 +1,0 @@
-# Gopkg.toml example
-#
-# Refer to https://golang.github.io/dep/docs/Gopkg.toml.html
-# for detailed Gopkg.toml documentation.
-#
-# required = ["github.com/user/thing/cmd/thing"]
-# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
-#
-# [[constraint]]
-#   name = "github.com/user/project"
-#   version = "1.0.0"
-#
-# [[constraint]]
-#   name = "github.com/user/project2"
-#   branch = "dev"
-#   source = "github.com/myfork/project2"
-#
-# [[override]]
-#   name = "github.com/x/y"
-#   version = "2.4.0"
-#
-# [prune]
-#   non-go = false
-#   go-tests = true
-#   unused-packages = true
-
-
-[prune]
-  go-tests = true
-  unused-packages = true
-
-[[constraint]]
-  name = "github.com/go-sql-driver/mysql"
-  version = "1.4.1"
-
-[[constraint]]
-  name = "github.com/jinzhu/gorm"
-  version = "1.0.0"
-
- [[constraint]]
-     name = "github.com/davecgh/go-spew"
-     version = "1.1.1"
-
-[[constraint]]
-      name = "github.com/DATA-DOG/go-sqlmock"
-      version = "1.3.3"
-
- [[constraint]]
-      name = "github.com/stretchr/testify"
-      version = "1.4.0"
