+++ pkg/usecase/presenter/user_presenter.go
@@ -1,6 +1,6 @@
 package presenter
 
-import "golang-clean-architecture/domain/model"
+import "golang-clean-architecture/pkg/domain/model"
 
 type UserPresenter interface {
 	ResponseUsers(u []*model.User) []*model.User
